Add Mocks in ReturnDroneFragmentDialogTest
