+++ app/src/androidTest/java/ch/epfl/sdp/AaaLocationWithoutPermissionTest.kt
@@ -7,7 +7,7 @@ import androidx.test.uiautomator.UiDevice
 import androidx.test.uiautomator.UiObject
 import androidx.test.uiautomator.UiSelector
 import ch.epfl.sdp.database.data_manager.MainDataManager
-import ch.epfl.sdp.drone.DroneInstanceMock
+import ch.epfl.sdp.ui.drone.DroneInstanceMock
 import ch.epfl.sdp.ui.MainActivity
 import ch.epfl.sdp.utils.CentralLocationManager
 import org.hamcrest.CoreMatchers.equalTo
