+++ app/src/androidTest/java/ch/epfl/sdp/drone/DroneUtilsTest.kt
@@ -1,5 +1,6 @@
 package ch.epfl.sdp.drone
 
+import ch.epfl.sdp.ui.drone.DroneInstanceMock
 import com.mapbox.mapboxsdk.geometry.LatLng
 import io.mavsdk.mission.Mission
 import org.junit.Assert
