+++ app/src/androidTest/java/ch/epfl/sdp/ui/MainActivityTest.kt
@@ -23,7 +23,7 @@ import androidx.test.rule.GrantPermissionRule.grant
 import androidx.test.uiautomator.UiDevice
 import ch.epfl.sdp.R
 import ch.epfl.sdp.database.data_manager.MainDataManager
-import ch.epfl.sdp.drone.DroneInstanceMock
+import ch.epfl.sdp.ui.drone.DroneInstanceMock
 import ch.epfl.sdp.ui.settings.SettingsActivity
 import ch.epfl.sdp.utils.Auth
 import org.hamcrest.CoreMatchers.*
