+++ app/src/androidTest/java/ch/epfl/sdp/ui/drone/DroneErrorTest.kt
@@ -1,7 +1,9 @@
-package ch.epfl.sdp.drone
+package ch.epfl.sdp.ui.drone
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.internal.runner.junit4.statement.UiThreadStatement.runOnUiThread
+import ch.epfl.sdp.drone.Drone
+import ch.epfl.sdp.drone.DroneUtils
 import com.mapbox.mapboxsdk.geometry.LatLng
 import io.mavsdk.telemetry.Telemetry
 import io.reactivex.Completable
