+++ app/src/androidTest/java/ch/epfl/sdp/ui/drone/ReturnDroneDialogFragmentTest.kt
@@ -11,7 +11,6 @@ import androidx.test.espresso.intent.rule.IntentsTestRule
 import androidx.test.espresso.matcher.RootMatchers.withDecorView
 import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.internal.runner.junit4.statement.UiThreadStatement
 import androidx.test.internal.runner.junit4.statement.UiThreadStatement.runOnUiThread
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.rule.GrantPermissionRule
@@ -21,14 +20,26 @@ import androidx.test.uiautomator.Until
 import ch.epfl.sdp.MainApplication
 import ch.epfl.sdp.MainApplication.Companion.applicationContext
 import ch.epfl.sdp.R
+import ch.epfl.sdp.database.dao.MockGroupDao
+import ch.epfl.sdp.database.dao.MockHeatmapDao
+import ch.epfl.sdp.database.dao.MockMarkerDao
+import ch.epfl.sdp.database.dao.MockUserDao
 import ch.epfl.sdp.database.data.Role
 import ch.epfl.sdp.database.data_manager.MainDataManager
+import ch.epfl.sdp.database.providers.HeatmapRepositoryProvider
+import ch.epfl.sdp.database.providers.MarkerRepositoryProvider
+import ch.epfl.sdp.database.providers.SearchGroupRepositoryProvider
+import ch.epfl.sdp.database.providers.UserRepositoryProvider
+import ch.epfl.sdp.database.repository.HeatmapRepository
+import ch.epfl.sdp.database.repository.MarkerRepository
+import ch.epfl.sdp.database.repository.SearchGroupRepository
+import ch.epfl.sdp.database.repository.UserRepository
 import ch.epfl.sdp.drone.Drone
 import ch.epfl.sdp.ui.maps.MapActivity
 import ch.epfl.sdp.utils.Auth
 import ch.epfl.sdp.utils.CentralLocationManager
 import com.mapbox.mapboxsdk.geometry.LatLng
-import io.mavsdk.telemetry.Telemetry.Position
+import io.mavsdk.telemetry.Telemetry
 import org.hamcrest.CoreMatchers.not
 import org.hamcrest.Matchers
 import org.hamcrest.Matchers.closeTo
@@ -69,9 +80,22 @@ class ReturnDroneDialogFragmentTest {
     @Throws(Exception::class)
     fun before() {
         MainDataManager.goOffline()
+        DroneInstanceMock.setupDefaultMocks()
+
+        // Do not use the real database, only use the offline version on the device
+        // Firebase.database.goOffline()
+        SearchGroupRepository.daoProvider = { MockGroupDao() }
+        HeatmapRepository.daoProvider = { MockHeatmapDao() }
+        MarkerRepository.daoProvider = { MockMarkerDao() }
+        UserRepository.daoProvider = { MockUserDao() }
+
+        SearchGroupRepositoryProvider.provide = { SearchGroupRepository() }
+        HeatmapRepositoryProvider.provide = { HeatmapRepository() }
+        MarkerRepositoryProvider.provide = { MarkerRepository() }
+        UserRepositoryProvider.provide = { UserRepository() }
 
         //Fake login
-        UiThreadStatement.runOnUiThread {
+        runOnUiThread {
             Auth.accountId.value = FAKE_ACCOUNT_ID
             Auth.loggedIn.value = true
         }
@@ -134,7 +158,7 @@ class ReturnDroneDialogFragmentTest {
     }
 
     @Test
-    fun testClickOnNeutralButtonShowsHomeSuccessToastWhenUserPositionIsNullButHomeIsNot() {
+    fun testClickOnNeutralButtonShowsHomeSuccessToastWhenUserPositionIsNull() {
         // Launch activity
         mActivityRule.launchActivity(intentWithGroupAndOperator)
         mUiDevice.wait(Until.hasObject(By.desc(applicationContext().getString(R.string.map_ready))), MAP_LOADING_TIMEOUT)
@@ -144,7 +168,7 @@ class ReturnDroneDialogFragmentTest {
         ReturnDroneDialogFragment().show(mActivityRule.activity.supportFragmentManager, mActivityRule.activity.getString(R.string.ReturnDroneDialogFragment))
 
         runOnUiThread {
-            Drone.homeLocationLiveData.value = Position(0.0, 0.0, 0.0f, 0.0f)
+            Drone.homeLocationLiveData.value = Telemetry.Position(0.0, 0.0, 0.0f, 0.0f)
         }
         CentralLocationManager.currentUserPosition.value = null
 
@@ -157,8 +181,6 @@ class ReturnDroneDialogFragmentTest {
         onView(withText(applicationContext().getString(R.string.drone_home_success)))
                 .inRoot(withDecorView(not(mActivityRule.activity.window.decorView)))
                 .check(matches(isDisplayed()))
-
-        Drone.homeLocationLiveData.value = null
     }
 
     @Test
