+++ app/src/androidTest/java/ch/epfl/sdp/ui/home/LoginNavFragmentTest.kt
@@ -20,7 +20,7 @@ import androidx.test.rule.GrantPermissionRule.grant
 import androidx.test.uiautomator.UiDevice
 import ch.epfl.sdp.MainApplication
 import ch.epfl.sdp.R
-import ch.epfl.sdp.drone.DroneInstanceMock
+import ch.epfl.sdp.ui.drone.DroneInstanceMock
 import ch.epfl.sdp.ui.MainActivity
 import ch.epfl.sdp.utils.Auth
 import com.google.android.gms.auth.api.signin.GoogleSignIn
