+++ app/src/androidTest/java/ch/epfl/sdp/ui/maps/MapActivityTest.kt
@@ -38,7 +38,7 @@ import ch.epfl.sdp.database.repository.MarkerRepository
 import ch.epfl.sdp.database.repository.SearchGroupRepository
 import ch.epfl.sdp.database.repository.UserRepository
 import ch.epfl.sdp.drone.Drone
-import ch.epfl.sdp.drone.DroneInstanceMock
+import ch.epfl.sdp.ui.drone.DroneInstanceMock
 import ch.epfl.sdp.mission.SimpleQuadStrategy
 import ch.epfl.sdp.mission.SpiralStrategy
 import ch.epfl.sdp.searcharea.QuadrilateralArea
