+++ app/src/main/res/layout/activity_map.xml
@@ -65,21 +65,6 @@
             app:fab_colorPressed="@color/colorLightGray"
             app:fab_icon="@drawable/ic_spiralstrat"
             app:fab_size="normal" />
-        <com.getbase.floatingactionbutton.FloatingActionButton
-            android:id="@+id/return_home_or_user"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="returnHomeOrUser"
-            app:fab_icon="@drawable/ic_return"
-            style="@style/FloatingButtonTheme"/>
-
-        <com.getbase.floatingactionbutton.FloatingActionButton
-            android:id="@+id/start_or_pause_button"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="startOrPauseMissionButton"
-            app:fab_icon="@drawable/ic_start"
-            style="@style/FloatingButtonTheme"/>
 
         <com.getbase.floatingactionbutton.FloatingActionsMenu
             android:id="@+id/floating_menu_button"
@@ -93,14 +78,20 @@
             mapbox:layout_constraintEnd_toEndOf="parent">
 
             <com.getbase.floatingactionbutton.FloatingActionButton
-                android:id="@+id/start_or_return_button"
-                style="@style/FloatingButtonTheme"
+                android:id="@+id/return_home_or_user"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:onClick="startMissionOrReturnHome"
-                app:fab_colorNormal="@color/colorWhite"
-                app:fab_colorPressed="@color/colorLightGray"
-                app:fab_icon="@drawable/ic_start" />
+                android:onClick="returnHomeOrUser"
+                app:fab_icon="@drawable/ic_return"
+                style="@style/FloatingButtonTheme"/>
+
+            <com.getbase.floatingactionbutton.FloatingActionButton
+                android:id="@+id/start_or_pause_button"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:onClick="startOrPauseMissionButton"
+                app:fab_icon="@drawable/ic_start"
+                style="@style/FloatingButtonTheme"/>
 
             <com.getbase.floatingactionbutton.FloatingActionButton
                 android:id="@+id/clear_button"
